restructure to resemble expected layout
