+++ Cargo.toml
@@ -18,6 +18,7 @@ crate-type = ["cdylib", "rlib"]
 [features]
 # for more explicit tests, cargo test --features=backtraces
 backtraces = ["cosmwasm-std/backtraces"]
+library = []
 
 [dependencies]
 cosmwasm-std = { version = "1.1" }
@@ -40,7 +41,7 @@ lazy_static = "1.4.0"
 #abstract-sdk = { version = "0.1.0-rc.3" }
 #
 
-abstract-os = { path = "../../Abstract/contracts/packages/abstract-os" }
+# abstract-os = { path = "../../Abstract/contracts/packages/abstract-os" }
 abstract-add-on = { path = "../../Abstract/contracts/packages/abstract-add-on", features = ["library"] }
 abstract-sdk = { path = "../../Abstract/contracts/packages/abstract-sdk" }
 #ibc-client = {path = "../../Abstract/contracts/contracts/native/ibc-client", features = ["library"]  }
