+++ examples/schema.rs
@@ -1,10 +1,9 @@
 use std::env::current_dir;
 use std::fs::create_dir_all;
 
-use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
 use abstract_addon_template::contract::TemplateAddOn;
 use abstract_addon_template::package::ConfigResponse;
-
+use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
 
 fn main() {
     let mut out_dir = current_dir().unwrap();
