+++ src/contract.rs
@@ -1,27 +1,20 @@
-use std::vec;
-
 use abstract_add_on::AddOnContract;
-use abstract_os::add_on::{ExecuteMsg, InstantiateMsg, MigrateMsg, QueryMsg};
-use abstract_os::objects::fee::Fee;
-use abstract_os::objects::AssetEntry;
-use abstract_os::EXCHANGE;
-use abstract_sdk::{ExecuteEndpoint, InstantiateEndpoint, MigrateEndpoint, QueryEndpoint};
-use cosmwasm_std::{
-    entry_point, to_binary, Addr, Binary, Deps, DepsMut, Env, MessageInfo, Order, Reply, ReplyOn,
-    Response, StdError, StdResult, SubMsg, WasmMsg,
+use abstract_sdk::abstract_os::add_on::{ExecuteMsg, InstantiateMsg, MigrateMsg, QueryMsg};
+use abstract_sdk::{
+    ExecuteEndpoint, InstantiateEndpoint, MigrateEndpoint, QueryEndpoint, ReplyEndpoint, EXCHANGE,
 };
-use cw2::{get_contract_version, set_contract_version};
-use cw20::{Balance, Cw20ReceiveMsg, MinterResponse};
-use protobuf::Message;
-use semver::Version;
+use cosmwasm_std::{
+    entry_point, Binary, Deps, DepsMut, Env, MessageInfo, Reply, Response, StdResult,
+};
 
 use crate::error::TemplateError;
-use crate::handlers::{commands, queries};
-use crate::package::state::{Config, CONFIG, COUNTS};
-use crate::package::{
+use crate::handlers::{self};
+
+use crate::msg::{
     TemplateExecuteMsg, TemplateInstantiateMsg, TemplateMigrateMsg, TemplateQueryMsg,
 };
 
+// As an add-on writer, the only changes necessary to this file are with the handlers and API dependencies on the `TEMPLATE_ADDON` const.
 pub type TemplateAddOn<'a> = AddOnContract<
     TemplateError,
     TemplateExecuteMsg,
@@ -33,79 +26,54 @@ pub type TemplateAddOn<'a> = AddOnContract<
 pub type TemplateResult = Result<Response, TemplateError>;
 pub const CONTRACT_NAME: &str = "template";
 
-/// Initial instantiation of the contract
-pub fn instantiate_handler(
-    mut deps: DepsMut,
-    _env: Env,
-    _info: MessageInfo,
-    _add_on: TemplateAddOn,
-    msg: TemplateInstantiateMsg,
-) -> TemplateResult {
-    // Initial config
-    let config: Config = Config {
-        max_count: msg.max_count,
-    };
+const CONTRACT_VERSION: &str = env!("CARGO_PKG_VERSION");
 
-    CONFIG.save(deps.storage, &config)?;
+/// Used as the foundation for building your addon. All entrypoints are executed through this const.
+/// `instantiate`, `query`, `execute`, `migrate`: entrypoint handlers
+/// `dependencies`:  Vec(`namespace:contract_name`) of Abstract API dependencies
+const TEMPLATE_ADDON: TemplateAddOn = TemplateAddOn::new(CONTRACT_NAME, CONTRACT_VERSION)
+    .with_instantiate(handlers::instantiate_handler)
+    .with_query(handlers::query_handler)
+    .with_execute(handlers::execute_handler)
+    .with_migrate(handlers::migrate_handler)
+    .with_dependencies(&[EXCHANGE]);
 
-    if let Some(initial_counts) = msg.initial_counts {
-        for (addr, count) in initial_counts {
-            let addr = deps.api.addr_validate(addr.as_str())?;
-
-            COUNTS.save(deps.storage, &addr, &count)?;
-        }
-    }
-
-    Ok(Response::new()
-        .add_attribute("action", "instantiate")
-        .add_attribute("contract", CONTRACT_NAME))
-}
-
-pub fn query_handler(
-    deps: Deps,
-    env: Env,
-    _add_on: &TemplateAddOn,
-    msg: TemplateQueryMsg,
-) -> StdResult<Binary> {
-    match msg {
-        TemplateQueryMsg::Config {} => to_binary(&queries::query_config(deps, env)?),
-        TemplateQueryMsg::UserCounts { users } => {
-            todo!()
-            // to_binary(&queries::query_user_counts(deps, env, users)?)
-        }
-        TemplateQueryMsg::UserCountList {
-            page_token,
-            page_size,
-        } => todo!(), // } => to_binary(&queries::query_user_count_list(deps, env)?),
-        TemplateQueryMsg::UserCount { user } => to_binary(&queries::query_count(deps, env, user)?),
-    }
-}
-
-/// Handle the `TemplateExecuteMsg`s sent to this add-on.
-pub fn execute_handler(
+/// Instantiate entrypoint
+#[cfg_attr(not(feature = "library"), entry_point)]
+pub fn instantiate(
     deps: DepsMut,
-    _env: Env,
+    env: Env,
     info: MessageInfo,
-    add_on: TemplateAddOn,
-    msg: TemplateExecuteMsg,
+    msg: InstantiateMsg<TemplateInstantiateMsg>,
 ) -> TemplateResult {
-    match msg {
-        TemplateExecuteMsg::UpdateConfig { max_count } => {
-            commands::update_config(deps, info, add_on, max_count)
-        }
-        TemplateExecuteMsg::Increment {} => {
-            commands::increment_sender(deps, info, add_on)
-        }
-    }
+    TEMPLATE_ADDON.instantiate(deps, env, info, msg)
 }
 
-/// Unused for now but provided here as an example
-/// Contract version is migrated automatically
-pub fn migrate_handler(
-    _deps: DepsMut,
-    _env: Env,
-    _add_on: TemplateAddOn,
-    _msg: TemplateMigrateMsg,
+/// Execute entrypoint
+#[cfg_attr(not(feature = "library"), entry_point)]
+pub fn execute(
+    deps: DepsMut,
+    env: Env,
+    info: MessageInfo,
+    msg: ExecuteMsg<TemplateExecuteMsg>,
 ) -> TemplateResult {
-    Ok(Response::default())
+    TEMPLATE_ADDON.execute(deps, env, info, msg)
+}
+
+/// Query entrypoint
+#[cfg_attr(not(feature = "library"), entry_point)]
+pub fn query(deps: Deps, env: Env, msg: QueryMsg<TemplateQueryMsg>) -> StdResult<Binary> {
+    TEMPLATE_ADDON.query(deps, env, msg)
+}
+
+/// Migrate entrypoint
+#[cfg_attr(not(feature = "library"), entry_point)]
+pub fn migrate(deps: DepsMut, env: Env, msg: MigrateMsg<TemplateMigrateMsg>) -> TemplateResult {
+    TEMPLATE_ADDON.migrate(deps, env, msg)
+}
+
+// Reply entrypoint
+#[cfg_attr(not(feature = "library"), entry_point)]
+pub fn reply(deps: DepsMut, env: Env, msg: Reply) -> TemplateResult {
+    TEMPLATE_ADDON.reply(deps, env, msg)
 }
