+++ src/handlers/mod.rs
@@ -1,3 +1,10 @@
 pub mod commands;
+pub mod instantiate;
+pub mod migrate;
 pub mod queries;
-pub mod replies;
+pub mod reply;
+
+pub use crate::handlers::{
+    commands::execute_handler, instantiate::instantiate_handler, migrate::migrate_handler,
+    queries::query_handler,reply::*
+};
