--- src/handlers/replies.rs
@@ -1,1 +1,0 @@
-const INSTANTIATE_REPLY_ID: u8 = 1u8;
