--- src/package/mod.rs
@@ -1,4 +1,0 @@
-pub mod msg;
-pub mod state;
-
-pub use msg::*;
